requirements.txt
yfinance==0.2.54
pandas==2.2.3
alpaca-py==0.35.0
apscheduler==3.10.4
pandas-market-calendars==4.4.1
requests==2.32.3
python-dotenv==1.0.1
lxml>=5.0.0
